Gardening: - Split classes into their own files - Replace Minitest with RSpec - Replace ThreadSafe::Set with Concurrent::Set (add dep on concurrent-ruby) - Set Travis to test on Ruby 2.5 to 3.0
