+++ .travis.yml
@@ -1,5 +1,12 @@
 sudo: false
 language: ruby
+cache: bundler
+
 rvm:
-  - 2.1.6
-before_install: gem install bundler -v 1.12.5
+  - 3.0.0
+  - 2.7.2
+  - 2.6.6
+  - 2.5.8
+
+matrix:
+  fast_finish: true
