+++ bin/console
@@ -1,14 +1,7 @@
 #!/usr/bin/env ruby
 
-require "bundler/setup"
-require "forky"
+require 'bundler/setup'
+require 'forky'
 
-# You can add fixtures and/or initialization code here to make experimenting
-# with your gem easier. You can also use a different console, if you like.
-
-# (If you use this, don't forget to add pry to your Gemfile!)
-# require "pry"
-# Pry.start
-
-require "irb"
-IRB.start
+require 'pry'
+Pry.start
