+++ forky.gemspec
@@ -4,30 +4,24 @@ $LOAD_PATH.unshift(lib) unless $LOAD_PATH.include?(lib)
 require 'forky/version'
 
 Gem::Specification.new do |spec|
-  spec.name          = "forky"
+  spec.name          = 'forky'
   spec.version       = Forky::VERSION
-  spec.authors       = ["Josh Bodah"]
-  spec.email         = ["jb3689@yahoo.com"]
+  spec.authors       = ['Josh Bodah']
+  spec.email         = ['jb3689@yahoo.com']
 
-  spec.summary       = %q{a library for simple distributed tasks}
-  spec.homepage      = "https://github.com/jbodah/forky"
-  spec.license       = "MIT"
+  spec.summary       = %q{Library for simple distributed tasks}
+  spec.homepage      = 'https://github.com/jbodah/forky'
+  spec.license       = 'MIT'
 
-  # Prevent pushing this gem to RubyGems.org. To allow pushes either set the 'allowed_push_host'
-  # to allow pushing to a single host or delete this section to allow pushing to any host.
-  if spec.respond_to?(:metadata)
-    spec.metadata['allowed_push_host'] = "TODO: Set to 'http://mygemserver.com'"
-  else
-    raise "RubyGems 2.0 or newer is required to protect against public gem pushes."
-  end
+  spec.files         = Dir.glob('lib/**/*') + %w[CHANGELOG.md LICENSE.txt README.md Rakefile]
+  spec.test_files    = Dir.glob('{test,spec}/**/*')
+  spec.require_path  = 'lib'
 
-  spec.files         = `git ls-files -z`.split("\x0").reject { |f| f.match(%r{^(test|spec|features)/}) }
-  spec.bindir        = "exe"
-  spec.executables   = spec.files.grep(%r{^exe/}) { |f| File.basename(f) }
-  spec.require_paths = ["lib"]
+  spec.add_dependency 'concurrent-ruby'
 
-  spec.add_development_dependency "bundler", "~> 1.12"
-  spec.add_development_dependency "rake", "~> 10.0"
-  spec.add_development_dependency "minitest", "~> 5.0"
-  spec.add_development_dependency "pry"
+  spec.add_development_dependency 'bundler'
+  spec.add_development_dependency 'pry'
+  spec.add_development_dependency 'rake'
+  spec.add_development_dependency 'rspec'
+  spec.add_development_dependency 'rubocop'
 end
