+++ lib/forky/mixin.rb
@@ -1,0 +1,7 @@
+module Forky
+  module Mixin
+    def async
+      AsyncProxy.new(self)
+    end
+  end
+end
