+++ spec/spec_helper.rb
@@ -1,0 +1,3 @@
+require 'bundler'
+Bundler.require
+require 'rspec'
