--- test/forky_test.rb
@@ -1,11 +1,0 @@
-require 'test_helper'
-
-class ForkyTest < Minitest::Test
-  def test_that_it_has_a_version_number
-    refute_nil ::Forky::VERSION
-  end
-
-  def test_it_does_something_useful
-    assert false
-  end
-end
