--- test/test_helper.rb
@@ -1,4 +1,0 @@
-$LOAD_PATH.unshift File.expand_path('../../lib', __FILE__)
-require 'forky'
-
-require 'minitest/autorun'
